Create 2_SimpleSums
2. Simple Sums
There are 15 marks available for this exercise.

Set up a new project and follow the below steps:

Declare a variable with data type integer. Initialise with a value of 8.
Declare a variable with data type integer. Initialise with a value of 10.
Declare a third variable of data type integer that adds the previous variables together.
Print the value of the sum to the console.