Create 4_PrimitiveQuiz
You have been asked to write a simple question and answer programme. There is only one question to answer: “What is the capital of France”.

The programme should ask the user this question, then prompt for a response.
The user then types in an answer.
If the answer is correct, the programme should tell the user that their answer was correct.
If the answer is incorrect, the programme should tell the user that their answer was wrong.
Advanced Requirements
What about capital letters? (e.g. paris vs Paris vs PaRis)
Add additional questions. (e.g. 10 Capitals Cities of Europe Quiz)