Update and rename 8_SimpleSearch.cpp to 08_SimpleSearch.cpp
Write a simple search engine that tests if a string is present in a string array. The array you need to declare should be initialised with the following values "Jake" "Zac", "Ian", "Ron", "Sam", "Dave". The string you are searching for is "Sam"

Your program should loop the array and test each array index against the search term. If the search term is found the program should output an appropriate message

Optional Requirements
Allow the user to enter the name to be searched